+++ eras/babbage/impl/cardano-ledger-babbage.cabal
@@ -80,7 +80,10 @@ library
         validation-selective
 
 library testlib
-    exposed-modules:  Test.Cardano.Ledger.Babbage.Arbitrary
+    exposed-modules:
+        Test.Cardano.Ledger.Babbage.Arbitrary
+        Test.Cardano.Ledger.Babbage.ImpTest
+
     visibility:       public
     hs-source-dirs:   testlib
     default-language: Haskell2010
@@ -92,9 +95,12 @@ library testlib
     build-depends:
         base,
         cardano-ledger-alonzo:testlib,
+        cardano-ledger-shelley:testlib,
+        cardano-crypto-class,
         cardano-ledger-babbage,
         cardano-ledger-binary,
         cardano-ledger-core,
+        cardano-ledger-alonzo:testlib,
         small-steps,
         QuickCheck
 
@@ -113,6 +119,7 @@ test-suite tests
         base,
         cardano-ledger-alonzo,
         cardano-ledger-babbage,
+        cardano-ledger-shelley:testlib,
         cardano-ledger-core:testlib,
         cardano-ledger-alonzo:testlib,
         data-default-class,
