+++ eras/babbage/impl/testlib/Test/Cardano/Ledger/Babbage/ImpTest.hs
@@ -1,0 +1,26 @@
+{-# LANGUAGE FlexibleContexts #-}
+{-# LANGUAGE UndecidableInstances #-}
+{-# OPTIONS_GHC -Wno-orphans #-}
+
+module Test.Cardano.Ledger.Babbage.ImpTest () where
+
+import Cardano.Crypto.DSIGN.Class (DSIGNAlgorithm (..))
+import Cardano.Crypto.Hash (Hash)
+import Cardano.Ledger.Babbage (BabbageEra)
+import Cardano.Ledger.Core (EraIndependentTxBody)
+import Cardano.Ledger.Crypto (Crypto (..))
+import Test.Cardano.Ledger.Alonzo.ImpTest (emptyAlonzoImpNES)
+import Test.Cardano.Ledger.Shelley.ImpTest (
+  EraImpTest (..),
+  shelleyImpWitsVKeyNeeded,
+ )
+
+instance
+  ( Crypto c
+  , Signable (DSIGN c) (Hash (HASH c) EraIndependentTxBody)
+  ) =>
+  EraImpTest (BabbageEra c)
+  where
+  emptyImpNES = emptyAlonzoImpNES
+
+  impWitsVKeyNeeded = shelleyImpWitsVKeyNeeded
