+++ eras/conway/impl/cardano-ledger-conway.cabal
@@ -68,6 +68,7 @@ library
     build-depends:
         base >=4.14 && <4.19,
         aeson,
+        data-default-class,
         cardano-crypto-class,
         cardano-data,
         cardano-ledger-binary >=1.1.3,
@@ -80,7 +81,6 @@ library
         cardano-slotting,
         cardano-strict-containers,
         containers,
-        data-default-class,
         deepseq,
         microlens,
         nothunks,
@@ -95,8 +95,8 @@ library testlib
     exposed-modules:
         Test.Cardano.Ledger.Conway.Arbitrary
         Test.Cardano.Ledger.Conway.Binary.RoundTrip
-        Test.Cardano.Ledger.Conway.ImpTest
         Test.Cardano.Ledger.Conway.PParamsSpec
+        Test.Cardano.Ledger.Conway.ImpTest
 
     visibility:       public
     hs-source-dirs:   testlib
@@ -108,29 +108,19 @@ library testlib
 
     build-depends:
         base,
-        transformers,
-        bytestring,
         cardano-data:testlib,
-        data-default-class,
         containers,
         microlens,
-        cardano-ledger-binary:{cardano-ledger-binary, testlib},
+        cardano-ledger-binary,
         cardano-crypto-class,
         cardano-ledger-core:{cardano-ledger-core, testlib},
         cardano-ledger-babbage:testlib,
         cardano-ledger-alonzo:{cardano-ledger-alonzo, testlib},
-        cardano-slotting,
         cardano-ledger-conway,
-        small-steps-test,
+        data-default-class,
         cardano-ledger-shelley,
-        cardano-ledger-babbage,
-        mtl,
-        time,
         cardano-strict-containers,
-        small-steps,
-        microlens-mtl,
-        unliftio,
-        deepseq
+        small-steps
 
 test-suite tests
     type:             exitcode-stdio-1.0
@@ -157,13 +147,12 @@ test-suite tests
         base,
         cardano-ledger-core:testlib,
         cardano-ledger-conway,
-        cardano-ledger-shelley,
+        cardano-ledger-shelley:{cardano-ledger-shelley, testlib},
         cardano-ledger-core,
         cardano-ledger-binary:testlib,
         containers,
         data-default-class,
         microlens,
         testlib,
-        text,
         cardano-strict-containers,
         microlens
