+++ eras/mary/impl/cardano-ledger-mary.cabal
@@ -1,6 +1,6 @@
 cabal-version:      3.0
 name:               cardano-ledger-mary
-version:            1.3.4.1
+version:            1.3.5.0
 license:            Apache-2.0
 maintainer:         operations@iohk.io
 author:             IOHK
@@ -75,7 +75,10 @@ library
         ghc-options: -fno-ignore-asserts
 
 library testlib
-    exposed-modules:  Test.Cardano.Ledger.Mary.Arbitrary
+    exposed-modules:
+        Test.Cardano.Ledger.Mary.Arbitrary
+        Test.Cardano.Ledger.Mary.ImpTest
+
     visibility:       public
     hs-source-dirs:   testlib
     default-language: Haskell2010
@@ -93,6 +96,7 @@ library testlib
         cardano-ledger-core,
         cardano-ledger-mary >=1.0.0.1,
         cardano-ledger-allegra:testlib,
+        cardano-ledger-shelley:testlib,
         cardano-strict-containers,
         QuickCheck
