+++ eras/mary/impl/test/Main.hs
@@ -1,8 +1,14 @@
+{-# LANGUAGE TypeApplications #-}
+
 module Main where
 
+import Cardano.Ledger.Mary (Mary)
+import Data.Data (Proxy (..))
 import Test.Cardano.Ledger.Common
 import qualified Test.Cardano.Ledger.Mary.BinarySpec as BinarySpec
+import Test.Cardano.Ledger.Mary.ImpTest ()
 import qualified Test.Cardano.Ledger.Mary.ValueSpec as ValueSpec
+import qualified Test.Cardano.Ledger.Shelley.ImpTestSpec as ImpTestSpec
 
 main :: IO ()
 main =
@@ -10,3 +16,4 @@ main =
     describe "Mary" $ do
       ValueSpec.spec
       BinarySpec.spec
+      ImpTestSpec.spec $ Proxy @Mary
