+++ eras/shelley/impl/src/Cardano/Ledger/Shelley/Rules/Delegs.hs
@@ -71,6 +71,7 @@ import Cardano.Ledger.Shelley.TxBody (
  )
 import Cardano.Ledger.Shelley.TxCert (ShelleyEraTxCert, pattern DelegStakeTxCert)
 import Cardano.Ledger.Slot (SlotNo)
+import Cardano.Ledger.TreeDiff (ToExpr)
 import Cardano.Ledger.UMap (UMElem (..), UMap (..), UView (..), fromCompact)
 import qualified Cardano.Ledger.UMap as UM
 import Control.DeepSeq
@@ -135,6 +136,10 @@ deriving stock instance
 instance
   NFData (PredicateFailure (EraRule "DELPL" era)) =>
   NFData (ShelleyDelegsPredFailure era)
+
+instance
+  ToExpr (PredicateFailure (EraRule "DELPL" era)) =>
+  ToExpr (ShelleyDelegsPredFailure era)
 
 instance
   ( EraTx era
