+++ eras/shelley/impl/src/Cardano/Ledger/Shelley/Rules/Ledger.hs
@@ -58,6 +58,7 @@ import Cardano.Ledger.Shelley.Rules.Reports (showTxCerts, synopsisCoinMap)
 import Cardano.Ledger.Shelley.Rules.Utxo (UtxoEnv (..))
 import Cardano.Ledger.Shelley.Rules.Utxow (ShelleyUTXOW, ShelleyUtxowPredFailure)
 import Cardano.Ledger.Slot (EpochNo, SlotNo, epochInfoEpoch)
+import Cardano.Ledger.TreeDiff (ToExpr)
 import Cardano.Ledger.UMap (depositMap)
 import Control.DeepSeq (NFData (..))
 import Control.Monad.Reader (Reader)
@@ -158,6 +159,12 @@ instance
           a <- decCBOR
           pure (2, DelegsFailure a)
         k -> invalidKey k
+
+instance
+  ( ToExpr (PredicateFailure (EraRule "UTXOW" era))
+  , ToExpr (PredicateFailure (EraRule "DELEGS" era))
+  ) =>
+  ToExpr (ShelleyLedgerPredFailure era)
 
 epochFromSlot :: SlotNo -> Reader Globals EpochNo
 epochFromSlot slot = do
