+++ eras/shelley/impl/src/Cardano/Ledger/Shelley/Rules/Utxo.hs
@@ -89,6 +89,7 @@ import Cardano.Ledger.Shelley.TxBody (
  )
 import Cardano.Ledger.Shelley.UTxO (consumed, produced, txup)
 import Cardano.Ledger.Slot (SlotNo)
+import Cardano.Ledger.TreeDiff (ToExpr)
 import Cardano.Ledger.UTxO (
   EraUTxO,
   UTxO (..),
@@ -309,6 +310,13 @@ instance
           outs <- decCBOR
           pure (2, OutputBootAddrAttrsTooBig outs)
         k -> invalidKey k
+
+instance
+  ( ToExpr (PPUPPredFailure era)
+  , ToExpr (Value era)
+  , ToExpr (TxOut era)
+  ) =>
+  ToExpr (ShelleyUtxoPredFailure era)
 
 instance
   ( EraTx era
