+++ libs/cardano-ledger-core/testlib/Test/Cardano/Ledger/Core/Utils.hs
@@ -1,9 +1,27 @@
+{-# LANGUAGE ScopedTypeVariables #-}
+{-# LANGUAGE TypeApplications #-}
+
 module Test.Cardano.Ledger.Core.Utils (
   unsafeBoundRational,
   epsilonMaybeEq,
+  testGlobals,
+  mkDummySafeHash,
 )
 where
 
+import Cardano.Ledger.BaseTypes (
+  EpochSize (..),
+  Globals (..),
+  Network (..),
+  mkActiveSlotCoeff,
+ )
+import Cardano.Ledger.Crypto (Crypto (..))
+import Cardano.Ledger.SafeHash (SafeHash, unsafeMakeSafeHash)
+import Cardano.Slotting.EpochInfo (fixedEpochInfo)
+import Cardano.Slotting.Time (SystemStart (..), mkSlotLength)
+import Data.Data (Proxy)
+import Data.Time.Clock.POSIX (posixSecondsToUTCTime)
+import Test.Cardano.Ledger.Binary.Random (mkDummyHash)
 import Test.Cardano.Ledger.Common
 import Test.Cardano.Ledger.Core.Rational (unsafeBoundRational)
 
@@ -26,3 +44,23 @@ epsilonMaybeEq epsilon x y =
     (absx, absy) = (abs x, abs y)
     n = epsilon * (1 + max absx absy)
     diff = abs (y - x)
+
+testGlobals :: Globals
+testGlobals =
+  Globals
+    { epochInfo = fixedEpochInfo (EpochSize 100) (mkSlotLength 1)
+    , slotsPerKESPeriod = 20
+    , stabilityWindow = 33
+    , randomnessStabilisationWindow = 33
+    , securityParameter = 10
+    , maxKESEvo = 10
+    , quorum = 5
+    , maxMajorPV = maxBound
+    , maxLovelaceSupply = 45 * 1000 * 1000 * 1000 * 1000 * 1000
+    , activeSlotCoeff = mkActiveSlotCoeff . unsafeBoundRational $ 0.9
+    , networkId = Testnet
+    , systemStart = SystemStart $ posixSecondsToUTCTime 0
+    }
+
+mkDummySafeHash :: forall c a. Crypto c => Proxy c -> Int -> SafeHash c a
+mkDummySafeHash _ = unsafeMakeSafeHash . mkDummyHash @(HASH c)
